[Refactor] #238 로드맵 관련 API 엔드포인트 변경
REST 설계 원칙에 따라 Roadmap 하위 도메인 관련 API들은 /roadmap 하위 경로로 위치하도록 변경
